requirements.txt
# Core web framework
fastapi==0.129.0
uvicorn==0.40.0
python-dotenv==1.2.1
jinja2>=3.1.0

# LLM SDKs
openai==2.20.0
google-generativeai==0.8.6

# Code formatting
black>=23.0.0
autopep8>=2.0.0

# Code quality & static analysis (pylint + radon as per proposal)
pylint>=3.0.0
radon>=6.0.0

# Data validation
pydantic==2.12.5

# HTTP client
httpx==0.28.1
requests==2.32.5

# Additional dependencies
starlette==0.52.1
